requirements.txt
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
xgboost==2.1.1
scikit-learn==1.4.2
fastapi==0.114.2
uvicorn==0.30.6
python-dotenv==1.0.1
mlflow==2.16.2
pyyaml==6.0.2
requests==2.32.3
langchain==0.3.7
langchain-openai==0.2.7
langchain-community==0.3.7
openai>=1.51.0         # (LangChain 대신 직접 SDK 쓰고 싶을 때 선택)